Branch: MatrixPilot_NASA_challenge - misc changes of things I notice while merging IP to AUAV3 - removed returns which were the last line of a function - updated AUTHORS.TXT - updated a few comments - Moved HardwareProfile to the /MatrixPilot folder to match the UAV3 folder. This is proper placement because eventually USB and other Microchip libraries will also use it.
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_NASA_challenge@2384 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/flightplan-logo.c
@@ -20,17 +20,16 @@
 
 
 #include "defines.h"
-#include "stdlib.h"
 
 #if (FLIGHT_PLAN_TYPE == FP_LOGO)
 
 
 struct logoInstructionDef {
-	unsigned int cmd		:  6 ;
-	unsigned int do_fly		:  1 ;
-	unsigned int use_param	:  1 ;
-	unsigned int subcmd		:  8 ;
-	int arg		: 16 ;
+	uint16_t cmd		:  6 ;
+	uint16_t do_fly		:  1 ;
+	uint16_t use_param	:  1 ;
+	uint16_t subcmd		:  8 ;
+	int16_t arg					: 16 ;
 } ;
 
 #define PLANE				0
